Setup frontend and created user management page

+++ server/src/controllers/userController.ts
@@ -2,11 +2,13 @@ import mongoose from "mongoose";
 import User from "../models/User";
 import { ExpressProps } from "../types";
 import { errorResponse, successResponse } from "./responseController";
+import { removeMediaFromCloudinary, uploadMediaToCloudinary } from "../libs/cloudinary-media-handler";
 
 // <!-- Create User -->
 export const createUser: ExpressProps = async (req, res, next) => {
   try {
     const body = req.body;
+
     const isEmailExist = await User.findOne({ email: body.email });
     if (isEmailExist) return errorResponse(res, {
       statusCode: 409,
@@ -16,7 +18,11 @@ export const createUser: ExpressProps = async (req, res, next) => {
       }
     });
 
-    await User.create(body);
+    const avatar = req.file ?
+      (await uploadMediaToCloudinary(req.file)).secure_url
+      : "";
+
+    await User.create({ ...body, avatar });
 
     return successResponse(res, {
       statusCode: 201,
@@ -24,6 +30,48 @@ export const createUser: ExpressProps = async (req, res, next) => {
     });
   } catch (error) {
     next(error)
+  }
+};
+
+// <!-- Get all Users -->
+export const getAllUsers: ExpressProps = async (req, res, next) => {
+  try {
+    const search = req.query.search || '';
+    const page = Number(req.query.page) || 1;
+    const limit = Number(req.query.limit) || 10;
+    const role = req.query.role?.toString().toLowerCase() || '';
+
+    const searchRegExp = new RegExp('.*' + search + '.*', 'i');
+    const filter: any = {
+      $or: [
+        { name: { $regex: searchRegExp } },
+        { email: { $regex: searchRegExp } },
+      ]
+    };
+    if (role) filter.role = role;
+
+    const users = await User.find(filter, { password: 0 })
+      .sort({ createdAt: -1 })
+      .limit(limit)
+      .skip((page - 1) * limit);
+
+    const count = await User.find(filter).countDocuments();
+
+    return successResponse(res, {
+      statusCode: 200,
+      message: 'Users were returned',
+      data: {
+        users,
+        pagination: {
+          totalPage: Math.ceil(count / limit),
+          currentPage: page,
+          previousPage: page > 1 ? page - 1 : null,
+          nextPage: page < Math.ceil(count / limit) ? page + 1 : null,
+        },
+      },
+    });
+  } catch (error) {
+    next(error);
   }
 };
 
@@ -62,6 +110,9 @@ export const deleteUserById: ExpressProps = async (req, res, next) => {
   try {
     const id = req.params.id;
 
+    const user = await User.findById(id);
+    if (!user) return errorResponse(res, { statusCode: 404, message: 'User does not exist' });
+    if (user?.avatar) await removeMediaFromCloudinary(user.avatar);
     await User.findByIdAndDelete(id);
 
     return successResponse(res, {
@@ -73,49 +124,6 @@ export const deleteUserById: ExpressProps = async (req, res, next) => {
     if (error instanceof mongoose.Error) {
       return errorResponse(res, { statusCode: 400, message: 'Invalid user id' });
     };
-    next(error);
-  }
-};
-
-// <!-- Get all Users -->
-export const getAllUsers: ExpressProps = async (req, res, next) => {
-  try {
-    const search = req.query.search || '';
-    const page = Number(req.query.page) || 1;
-    const limit = Number(req.query.limit) || 10;
-    const order_by = req.query.order_by?.toString().toLowerCase() || '';
-
-    const searchRegExp = new RegExp('.*' + search + '.*', 'i');
-    const filter: any = {
-      isAdmin: { $ne: true },
-      $or: [
-        { name: { $regex: searchRegExp } },
-        { email: { $regex: searchRegExp } },
-      ]
-    };
-    if (order_by && order_by !== "all") filter.role = order_by;
-
-    const users = await User.find(filter, { password: 0 })
-      .sort({ createdAt: -1 })
-      .limit(limit)
-      .skip((page - 1) * limit);
-
-    const count = await User.find(filter).countDocuments();
-
-    return successResponse(res, {
-      statusCode: 200,
-      message: 'Users were returned',
-      data: {
-        users,
-        pagination: {
-          totalPage: Math.ceil(count / limit),
-          currentPage: page,
-          previousPage: page > 1 ? page - 1 : null,
-          nextPage: page < Math.ceil(count / limit) ? page + 1 : null,
-        },
-      },
-    });
-  } catch (error) {
     next(error);
   }
 };
